update permission check (wip)
